requirements.txt
streamlit==1.0.0
torch==1.8.1
transformers==4.6.0
pillow==8.2.0
diffusers==0.1.0
